LPS-102957 Move all display contexts to the new app, related to page template collections
